+++ main.py
@@ -1,9 +1,9 @@
 import sys
 import time
-import Patcher
+import patcher
 
 if __name__ == "__main__":
     start = time.time()
-    Patcher.patch(*sys.argv)
+    patcher.patch(*sys.argv)
     end = time.time()
     print(f"Patched in {end-start:.2f}s")
