+++ patcher/BasicBinaryParser.py
